Fix auto-version bumping based on PR labels (#1586)
## Description

Fixes version bumping in release drafter workflow. You can label a PR
with `major`, `minor`, or `patch` to indicate which component of the
version number the PR should increment (defaults to `patch` for
unlabeled PRs). The release drafter will tag & name the draft release
according to the maximum bumped version component of the set of PRs in
the release.

## Changes

- Uses the correct variable for tagging & describing the draft release
- Updates some permissions in the action

- [ ] I have reviewed the [Guidelines for Contributing](CONTRIBUTING.md)
and the [Code of Conduct](CODE_OF_CONDUCT.md).